Updated imports for cirular imports and fixed upload clean method according to failed tests
